+++ src/Symfony/Component/HttpFoundation/Session/Storage/NativeSessionStorage.php
@@ -12,6 +12,9 @@
 namespace Symfony\Component\HttpFoundation\Session\Storage;
 
 use Symfony\Component\HttpFoundation\Session\SessionBagInterface;
+use Symfony\Component\HttpFoundation\Session\Storage\Handler\StrictSessionHandler;
+use Symfony\Component\HttpFoundation\Session\Storage\Proxy\AbstractProxy;
+use Symfony\Component\HttpFoundation\Session\Storage\Proxy\SessionHandlerProxy;
 
 /**
  * This provides a base class for session attribute storage.
@@ -23,7 +26,7 @@ class NativeSessionStorage implements SessionStorageInterface
     /**
      * @var SessionBagInterface[]
      */
-    protected $bags;
+    protected $bags = array();
 
     /**
      * @var bool
@@ -36,7 +39,7 @@ class NativeSessionStorage implements SessionStorageInterface
     protected $closed = false;
 
     /**
-     * @var \SessionHandlerInterface
+     * @var AbstractProxy|\SessionHandlerInterface
      */
     protected $saveHandler;
 
@@ -85,13 +88,18 @@ class NativeSessionStorage implements SessionStorageInterface
      * sid_bits_per_character, "5"
      * trans_sid_hosts, $_SERVER['HTTP_HOST']
      * trans_sid_tags, "a=href,area=href,frame=src,form="
+     *
+     * @param array                         $options Session configuration options
+     * @param \SessionHandlerInterface|null $handler
+     * @param MetadataBag                   $metaBag MetadataBag
      */
-    public function __construct(array $options = array(), \SessionHandlerInterface $handler = null, MetadataBag $metaBag = null)
+    public function __construct(array $options = array(), $handler = null, MetadataBag $metaBag = null)
     {
         $options += array(
-            // disable by default because it's managed by HeaderBag (if used)
-            'cache_limiter' => '',
+            'cache_limiter' => 'private_no_expire',
             'use_cookies' => 1,
+            'lazy_write' => 1,
+            'use_strict_mode' => 1,
         );
 
         session_register_shutdown();
@@ -104,7 +112,7 @@ class NativeSessionStorage implements SessionStorageInterface
     /**
      * Gets the save handler instance.
      *
-     * @return \SessionHandlerInterface
+     * @return AbstractProxy|\SessionHandlerInterface
      */
     public function getSaveHandler()
     {
@@ -143,21 +151,15 @@ class NativeSessionStorage implements SessionStorageInterface
      */
     public function getId()
     {
-        return session_id();
+        return $this->saveHandler->getId();
     }
 
     /**
      * {@inheritdoc}
-     *
-     * @throws \LogicException When the session is active
      */
     public function setId($id)
     {
-        if (\PHP_SESSION_ACTIVE === session_status()) {
-            throw new \LogicException('Cannot change the ID of an active session');
-        }
-
-        session_id($id);
+        $this->saveHandler->setId($id);
     }
 
     /**
@@ -165,21 +167,15 @@ class NativeSessionStorage implements SessionStorageInterface
      */
     public function getName()
     {
-        return session_name();
+        return $this->saveHandler->getName();
     }
 
     /**
      * {@inheritdoc}
-     *
-     * @throws \LogicException When the session is active
      */
     public function setName($name)
     {
-        if (\PHP_SESSION_ACTIVE === session_status()) {
-            throw new \LogicException('Cannot change the name of an active session');
-        }
-
-        session_name($name);
+        $this->saveHandler->setName($name);
     }
 
     /**
@@ -218,20 +214,38 @@ class NativeSessionStorage implements SessionStorageInterface
      */
     public function save()
     {
+        $session = $_SESSION;
+
+        foreach ($this->bags as $bag) {
+            if (empty($_SESSION[$key = $bag->getStorageKey()])) {
+                unset($_SESSION[$key]);
+            }
+        }
+        if (array($key = $this->metadataBag->getStorageKey()) === array_keys($_SESSION)) {
+            unset($_SESSION[$key]);
+        }
+
         // Register custom error handler to catch a possible failure warning during session write
-        set_error_handler(function ($errno, $errstr, $errfile, $errline, $errcontext) {
-            throw new \ErrorException($errstr, $errno, E_WARNING, $errfile, $errline, $errcontext);
+        set_error_handler(function ($errno, $errstr, $errfile, $errline) {
+            throw new \ErrorException($errstr, $errno, E_WARNING, $errfile, $errline);
         }, E_WARNING);
 
         try {
+            $e = null;
             session_write_close();
-            restore_error_handler();
         } catch (\ErrorException $e) {
+        } finally {
+            restore_error_handler();
+            $_SESSION = $session;
+        }
+        if (null !== $e) {
             // The default PHP error message is not very helpful, as it does not give any information on the current save handler.
             // Therefore, we catch this error and trigger a warning with a better error message
             $handler = $this->getSaveHandler();
+            if ($handler instanceof SessionHandlerProxy) {
+                $handler = $handler->getHandler();
+            }
 
-            restore_error_handler();
             trigger_error(sprintf('session_write_close(): Failed to write session data with %s handler', get_class($handler)), E_USER_WARNING);
         }
 
@@ -258,8 +272,6 @@ class NativeSessionStorage implements SessionStorageInterface
 
     /**
      * {@inheritdoc}
-     *
-     * @throws \LogicException When the session is already started
      */
     public function registerBag(SessionBagInterface $bag)
     {
@@ -279,7 +291,7 @@ class NativeSessionStorage implements SessionStorageInterface
             throw new \InvalidArgumentException(sprintf('The SessionBagInterface %s is not registered.', $name));
         }
 
-        if (!$this->started && \PHP_SESSION_ACTIVE === session_status()) {
+        if (!$this->started && $this->saveHandler->isActive()) {
             $this->loadSession();
         } elseif (!$this->started) {
             $this->start();
@@ -372,16 +384,36 @@ class NativeSessionStorage implements SessionStorageInterface
      * @see http://php.net/sessionhandlerinterface
      * @see http://php.net/sessionhandler
      * @see http://github.com/drak/NativeSession
+     *
+     * @param \SessionHandlerInterface|null $saveHandler
+     *
+     * @throws \InvalidArgumentException
      */
-    public function setSaveHandler(\SessionHandlerInterface $saveHandler = null)
+    public function setSaveHandler($saveHandler = null)
     {
-        $this->saveHandler = $saveHandler ?: new \SessionHandler();
+        if (!$saveHandler instanceof AbstractProxy &&
+            !$saveHandler instanceof \SessionHandlerInterface &&
+            null !== $saveHandler) {
+            throw new \InvalidArgumentException('Must be instance of AbstractProxy; implement \SessionHandlerInterface; or be null.');
+        }
 
         if (headers_sent($file, $line)) {
             throw new \RuntimeException(sprintf('Failed to set the session handler because headers have already been sent by "%s" at line %d.', $file, $line));
         }
 
-        session_set_save_handler($this->saveHandler, false);
+        // Wrap $saveHandler in proxy and prevent double wrapping of proxy
+        if (!$saveHandler instanceof AbstractProxy && $saveHandler instanceof \SessionHandlerInterface) {
+            $saveHandler = new SessionHandlerProxy($saveHandler);
+        } elseif (!$saveHandler instanceof AbstractProxy) {
+            $saveHandler = new SessionHandlerProxy(new StrictSessionHandler(new \SessionHandler()));
+        }
+        $this->saveHandler = $saveHandler;
+
+        if ($this->saveHandler instanceof SessionHandlerProxy) {
+            session_set_save_handler($this->saveHandler->getHandler(), false);
+        } elseif ($this->saveHandler instanceof \SessionHandlerInterface) {
+            session_set_save_handler($this->saveHandler, false);
+        }
     }
 
     /**
@@ -400,12 +432,11 @@ class NativeSessionStorage implements SessionStorageInterface
             $session = &$_SESSION;
         }
 
-        $bags = $this->bags;
-        $bags[] = $this->metadataBag;
+        $bags = array_merge($this->bags, array($this->metadataBag));
 
         foreach ($bags as $bag) {
             $key = $bag->getStorageKey();
-            $session[$key] = $session[$key] ?? array();
+            $session[$key] = isset($session[$key]) ? $session[$key] : array();
             $bag->initialize($session[$key]);
         }